// src/types/IExchange.ts
export enum OrderSide {
    Buy = 'buy',
    Sell = 'sell',
  }
  
  export enum OrderType {
    Market = 'market',
    Limit = 'limit',
  }
  
  export interface Order {
    id: string
    market: string
    side: string
    type: string
    price: number
    quantity: number
    status?: string
    filled?: number
    remaining?: number
    createdAt: Date
    clientId?: string | null
  }